moving the encoder to its own file
