+++ app/src/main/java/com/example/clicker/presentation/selfStreaming/SelfStreamingFragment.kt
@@ -20,6 +20,7 @@ import android.media.ImageReader
 import android.media.MediaCodec
 import android.media.MediaCodecInfo
 import android.media.MediaFormat
+import android.media.MediaMuxer
 import android.os.Build
 import android.os.Bundle
 import android.os.Handler
@@ -600,121 +601,22 @@ class SelfStreamingFragment : Fragment() {
         return EncoderWrapper(
             width, height,
             RECORDER_VIDEO_BITRATE,
-           fps
+           fps,
+            orientation,
+            outputFile
         )
     }
 
 
-}
-/***
- * -------------------------------------------------ENCODER CLASS-------------------------------------------------------------------
- ******/
-class EncoderWrapper(
-    width: Int,
-    height: Int,
-    bitRate: Int,
-    frameRate: Int,
-
-){
-
-
-    companion object {
-        val TAG = "EncoderWrapper"
-
-        val IFRAME_INTERVAL = 1 // sync one frame every second
-        public const val VIDEO_CODEC_ID_HEVC: Int = 0
-        public const val VIDEO_CODEC_ID_H264: Int = 1
-        public const val VIDEO_CODEC_ID_AV1: Int = 2
-    }
-    /**
-     * Returns the encoder's input surface.
-     */
-    public fun getInputSurface(): Surface {
-        return mInputSurface
-    }
-    private val mInputSurface: Surface by lazy {
-            mEncoder!!.createInputSurface()
-    }
-
-    private val mVideoCodec = VIDEO_CODEC_ID_AV1
-
-    private val mMimeType = idToType(mVideoCodec)
-    private val dynamicRange =DynamicRangeProfiles.STANDARD
-
-    private val mEncoder: MediaCodec? by lazy {
-        MediaCodec.createEncoderByType("video/avc")
-    }
-
-    /**
-     * Configures encoder
-     */
-    init {
-
-            val codecProfile = when (mVideoCodec) {
-                VIDEO_CODEC_ID_HEVC -> when {
-                    dynamicRange == DynamicRangeProfiles.HLG10 ->
-                        MediaCodecInfo.CodecProfileLevel.HEVCProfileMain10
-                    dynamicRange == DynamicRangeProfiles.HDR10 ->
-                        MediaCodecInfo.CodecProfileLevel.HEVCProfileMain10HDR10
-                    dynamicRange == DynamicRangeProfiles.HDR10_PLUS ->
-                        MediaCodecInfo.CodecProfileLevel.HEVCProfileMain10HDR10Plus
-                    else -> -1
-                }
-                VIDEO_CODEC_ID_AV1 -> when {
-                    dynamicRange == DynamicRangeProfiles.HLG10 ->
-                        MediaCodecInfo.CodecProfileLevel.AV1ProfileMain10
-                    dynamicRange == DynamicRangeProfiles.HDR10 ->
-                        MediaCodecInfo.CodecProfileLevel.AV1ProfileMain10HDR10
-                    dynamicRange == DynamicRangeProfiles.HDR10_PLUS ->
-                        MediaCodecInfo.CodecProfileLevel.AV1ProfileMain10HDR10Plus
-                    else -> -1
-                }
-                else -> -1
-            }
-
-            val format = MediaFormat.createVideoFormat(mMimeType, width, height)
-
-            // Set some properties.  Failing to specify some of these can cause the MediaCodec
-            // configure() call to throw an unhelpful exception.
-            format.setInteger(MediaFormat.KEY_COLOR_FORMAT,
-                MediaCodecInfo.CodecCapabilities.COLOR_FormatSurface)
-            format.setInteger(MediaFormat.KEY_BIT_RATE, bitRate)
-            format.setInteger(MediaFormat.KEY_FRAME_RATE, frameRate)
-            format.setInteger(MediaFormat.KEY_I_FRAME_INTERVAL, IFRAME_INTERVAL)
-
-            if (codecProfile != -1) {
-                format.setInteger(MediaFormat.KEY_PROFILE, codecProfile)
-                format.setInteger(MediaFormat.KEY_COLOR_STANDARD, MediaFormat.COLOR_STANDARD_BT2020)
-                format.setInteger(MediaFormat.KEY_COLOR_RANGE, MediaFormat.COLOR_RANGE_FULL)
-                format.setInteger(MediaFormat.KEY_COLOR_TRANSFER, getTransferFunction())
-                format.setFeatureEnabled(MediaCodecInfo.CodecCapabilities.FEATURE_HdrEditing, true)
-            }
-
-           Log.d(TAG, "format: " + format)
-
-            // Create a MediaCodec encoder, and configure it with our format.  Get a Surface
-            // we can use for input and wrap it with a class that handles the EGL work.
-            mEncoder!!.configure(format, null, null, MediaCodec.CONFIGURE_FLAG_ENCODE)
-        }
 
 
 
-
-    private fun idToType(videoCodecId: Int): String = when (videoCodecId) {
-        VIDEO_CODEC_ID_H264 -> MediaFormat.MIMETYPE_VIDEO_AVC
-        VIDEO_CODEC_ID_HEVC -> MediaFormat.MIMETYPE_VIDEO_HEVC
-        VIDEO_CODEC_ID_AV1 -> MediaFormat.MIMETYPE_VIDEO_AV1
-        else -> throw RuntimeException("Unexpected video codec id " + videoCodecId)
-    }
-
-    private fun getTransferFunction() = when (dynamicRange) {
-        DynamicRangeProfiles.HLG10 -> MediaFormat.COLOR_TRANSFER_HLG
-        DynamicRangeProfiles.HDR10 -> MediaFormat.COLOR_TRANSFER_ST2084
-        DynamicRangeProfiles.HDR10_PLUS -> MediaFormat.COLOR_TRANSFER_ST2084
-        else -> MediaFormat.COLOR_TRANSFER_SDR_VIDEO
-    }
 
 }
+
+
+
+
